hdmi enabled EFI while logging in
I modified the boot option so that hdmi works while logging in with smbios13.2. Now you can plug in the hdmi cable and boot.
You can adjust the sound with F1, F2, and F3 keys, and you can adjust the screen brightness with Fn key, F11, and F12.
In the keyboard setting of the control panel, please specify the shortcut as F11 key for dark, F12 key for bright, and the mission control key as F10.
Good luck!